minor tweaking to code
I was going to clean up all my code completly, but I realized now that most of it isn't that bad considering the context. I think the biggest issue with the code is too much of the code is in PlayerController. I think I could solve this next time by avoding the term "controller" and instead breaking everything down to very specific scripts for the player. I do also kind of regret how unmodular my health UI script is, but it really didn't need to be modular considering the requirements of the project. Compared to 8-Bit Battles, the code to this project isn't too bad. I would say the biggest things I learned from this is
A) How to use the Animator better
B) IEnumerator for timers
C) Polymorphism in C# being used in a practical way

I think the biggest thing to improve from this is to make more classes and to make more methods so I can reuse code better.